+++ include/physics/BNS/body.h
@@ -90,7 +90,7 @@ public:
   double getInternalenergy() const{return internalenergy_;}
   point_t getLinMomentum() const { 
     point_t res = {};
-    for(int i = 0 ; i < dimension; ++i){
+    for(size_t i = 0 ; i < dimension; ++i){
       res[i] = velocity_[i] * mass_;
     }
     return res;
